+++ tests/test.js
@@ -111,7 +111,7 @@ describe('React unit/intergration testing', () => {
   test('can accept props', async () => {
     // jest-prop-type-error will throw if props dont match proptypes
     render(<ScrollUpButton 
-      ShowAtPostion={200}
+      ShowAtPosition={200}
       StopPosition={100}
       style={{}}
       ToggledStyle={{}}
@@ -185,7 +185,7 @@ describe('custom proptype validators', () => {
 
   test('LessThanShowAtPosition throws when StopPosition is greater than ShowAtPosition.', () => {
     expect(() => {
-      render(<ScrollUpButton StopPosition={100} ShowAtPostion={90} />)
+      render(<ScrollUpButton StopPosition={100} ShowAtPosition={90} />)
     }).toThrow()
   });